More stuff
Increased chances of a response to taylorswift as long as the message
isn't because of a URL or a YouTube link being posted.

I also decided to create a variable storing one of the word-triggered
responses so we don't clog up the chat with garbage (not that there's
really a problem with that, but...)

I also made it strip colors from messages. This allows us to parse
messages easier and lets us print the MOTD to screen when we connect
without it looking like trash.

We also respond to ".bots" now with a configurable message.